requirements.txt
html2text==2024.2.26
iniconfig==2.0.0
numpy==2.1.2
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.3
